some cleanup, remove unused
